fixes for both
