+++ .github/workflows/roundtrip/wait-and-test.sh
@@ -11,13 +11,16 @@ echo "[INFO] App version: ${app_version}"
 
 if [ $1 = backend ]; then
   VITE_PROXY='{"/api":"http://localhost:5432","/auth":"http://localhost:5432"}'
-  VITE_TDF_CFG='{oidc:{host:"http://localhost:65432/auth/realms/tdf",clientId:"browsertest"},kas:"http://localhost:65432/api/kas",reader:"https://secure.virtru.com/start?htmlProtocol=1"}'
+  VITE_TDF_CFG='{"oidc":{"host":"http://localhost:65432/auth/realms/tdf","clientId":"browsertest"},"kas":"http://localhost:65432/api/kas","reader":"https://secure.virtru.com/start?htmlProtocol=1"}'
 else
   VITE_PROXY='{"/kas":"http://localhost:8080","/auth":"http://localhost:8888"}'
-  VITE_TDF_CFG='{oidc:{host:"http://localhost:65432/auth/realms/opentdf",clientId:"browsertest"},kas:"http://localhost:65432/kas",reader:"https://secure.virtru.com/start?htmlProtocol=1"}'
+  VITE_TDF_CFG='{"oidc":{"host":"http://localhost:65432/auth/realms/opentdf","clientId":"browsertest"},"kas":"http://localhost:65432/kas","reader":"https://secure.virtru.com/start?htmlProtocol=1"}'
 fi
 export VITE_PROXY
 export VITE_TDF_CFG
+
+
+# VITE_PROXY='{"/api":"http://localhost:5432","/auth":"http://localhost:5432"}' VITE_TDF_CFG='{"oidc":{"host":"http://localhost:65432/auth/realms/tdf","clientId":"browsertest"},"kas":"http://localhost:65432/api/kas","reader":"https://secure.virtru.com/start?htmlProtocol=1"}' npm run dev
 
 _wait-for() {
   echo "[INFO] In retry loop for quickstarted opentdf backend..."
