+++ web-app/package.json
@@ -17,7 +17,7 @@
   "dependencies": {
     "@opentdf/client": "file:../lib/opentdf-client-2.0.0.tgz",
     "clsx": "^2.0.0",
-    "native-file-system-adapter": "^3.0.0",
+    "native-file-system-adapter": "^3.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
@@ -26,7 +26,7 @@
     "@rollup/plugin-inject": "^5.0.3",
     "@types/react": "^18.2.17",
     "@types/react-dom": "^18.2.7",
-    "@types/wicg-file-system-access": "^2020.9.6",
+    "@types/wicg-file-system-access": "^2023.10.5",
     "@typescript-eslint/eslint-plugin": "^6.2.1",
     "@typescript-eslint/parser": "^6.2.1",
     "@vitejs/plugin-react": "^4.0.4",
