+++ web-app/vite.config.ts
@@ -4,6 +4,20 @@ import react from '@vitejs/plugin-react';
 
 const require = createRequire(import.meta.url)
 
+function proxy(): Record<string, string> {
+  console.log(process.env);
+  const { VITE_PROXY } = process.env;
+  if (VITE_PROXY) {
+    console.log(`using VITE_PROXY [${VITE_PROXY}]`);
+    return JSON.parse(VITE_PROXY);
+  }
+  console.log("using standard VITE_PROXY");
+  return {
+    '/kas': 'http://localhost:8080',
+    '/auth': 'http://localhost:8888',
+  };
+}
+
 // https://vitejs.dev/config/
 export default defineConfig({
   build: {
@@ -14,9 +28,6 @@ export default defineConfig({
   plugins: [react()],
   server: {
     port: 65432,
-    proxy: {
-      '/kas': 'http://localhost:8080',
-      '/auth': 'http://localhost:8888',
-    },
+    proxy: proxy(),
   },
 });
